+++ ExportCategories.lua
@@ -488,8 +488,10 @@ DF.ExportCategories = {
         "roleIconHideDPS",
         "roleIconOnlyInCombat",
         "roleIconHideOnlyInCombat",
-        "roleIconHideInCombat",
-        
+        "roleIconExternalTank",
+        "roleIconExternalHealer",
+        "roleIconExternalDPS",
+
         -- Leader Icon
         "leaderIconEnabled",
         "leaderIconAnchor",